fix: :bug: add trackBy to initial ngFor to prevent DOM rerendering and collapsing of expansion panels.
Added a trackBy function to the ngFor iterating over all of the teams. This makes the ngFor not re-render literally everything (and thus collapse the open panels) when updating team data. Instead, it will simply add new teams, or remove teams that are missing. This also makes the opened panels not collapse since they aren't being remade.