fix: :bug: deleting member now removes member from list on team-info panel
deleting a member wasn't clearing the team-info panel, so I fixed that, then they were still showing in the team list on the same panel. Fixed that. There's probably a better way to do it (I thought the updateTeamDetails function would do it, but it's not for some reason), but it works for now.